Project_1/claude-code-skills-lab-main/task-management-api/pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Task Management API with FastAPI and SQLModel"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
